Photometry Class Rework Upload
There is now a class that can be used and that makes things A LOT easier. It is all set up so that you should be able to call and run it from an external source. It holds onto it's apertures and results table, and has a method you call to run it on a file, which should be done externally. It saves it's master tab regularly, and does not yet have methods to export/save the data in a more usuable format, although that is coming