Add export methods to photometry class
Added 3 export methods to the photometry class, one that unpacks and gives a table for each source, where the columns are now a type of result value, and one that unpacks and creates a table for each value type (so it has the same 'each column is a source' as the master table). The third export method collapses the dictionary stored in each cell to the most important value; the aperture_sum, and then returns that resulting table.